implement parking session ending and starting

+++ station_admin/models/user_profile.py
@@ -4,7 +4,7 @@ from django.contrib.auth.models import User
 
 class UserProfile(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
-    credits = models.FloatField(default=0)
+    credits = models.DecimalField(default=0, decimal_places=2, max_digits=10)
 
     def full_name(self) -> str:
         return f"{self.user.first_name} {self.user.last_name}"
