+++ station_admin/views/client/user/add_credits_view.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from django.http import HttpResponse, HttpRequest
 from django.shortcuts import redirect
 from django.views import View
@@ -9,7 +11,7 @@ class AddCreditsView(View):
         credit_amount = form_data['credit_amount']
 
         try:
-            credit_amount = float(credit_amount)
+            credit_amount = Decimal(credit_amount)
         except ValueError:
             messages.error(request, 'Credit amount must be a float number')
             return redirect('user_profile')
